Massive refactor of entire MCP client
