+++ hexstrike-ai-mcp.json
@@ -1,11 +1,13 @@
 {
   "mcpServers": {
     "hexstrike-ai": {
-      "command": "python3",
+      "command": "/path/to/hexstrike/hexstrike-env/bin/python",
       "args": [
-        "/path/hexstrike_mcp.py",
+        "/path/to/hexstrike/hexstrike_mcp.py",
         "--server",
-        "http://IPADDRESS:8888"
+        "http://127.0.0.1:8888",
+        "--profile", 
+        "full"
       ],
       "description": "HexStrike AI Community Edition - Advanced Cybersecurity Automation Platform. Turn off alwaysAllow if you dont want autonomous execution!",
       "timeout": 300,
