+++ mcp_core/colored_formatter.py
@@ -1,0 +1,26 @@
+import logging
+from mcp_core.hexstrikecolors import HexStrikeColors
+
+class ColoredFormatter(logging.Formatter):
+    """Enhanced formatter with colors and emojis for MCP client - matches server styling"""
+    COLORS = {
+        'DEBUG': HexStrikeColors.DEBUG,
+        'INFO': HexStrikeColors.SUCCESS,
+        'WARNING': HexStrikeColors.WARNING,
+        'ERROR': HexStrikeColors.ERROR,
+        'CRITICAL': HexStrikeColors.CRITICAL
+    }
+
+    EMOJIS = {
+        'DEBUG': '🔍',
+        'INFO': '✅',
+        'WARNING': '⚠️',
+        'ERROR': '❌',
+        'CRITICAL': '🔥'
+    }
+
+    def format(self, record):
+        emoji = self.EMOJIS.get(record.levelname, '📝')
+        color = self.COLORS.get(record.levelname, HexStrikeColors.BRIGHT_WHITE)
+        record.msg = f"{color}{emoji} {record.msg}{HexStrikeColors.RESET}"
+        return super().format(record)
