+++ mcp_tools/api_scan/jwt_analyzer.py
@@ -1,0 +1,44 @@
+# mcp_tools/api_scan/jwt_analyzer.py
+
+from typing import Dict, Any
+
+def register_jwt_analyzer_tool(mcp, hexstrike_client, logger):
+
+    @mcp.tool()
+    def jwt_analyzer(jwt_token: str, target_url: str = "") -> Dict[str, Any]:
+        """
+        Advanced JWT token analysis and vulnerability testing.
+
+        Args:
+            jwt_token: JWT token to analyze
+            target_url: Optional target URL for testing token manipulation
+
+        Returns:
+            JWT analysis results with vulnerability assessment and attack vectors
+        """
+        data = {
+            "jwt_token": jwt_token,
+            "target_url": target_url
+        }
+
+        logger.info(f"🔍 Starting JWT security analysis")
+        result = hexstrike_client.safe_post("api/tools/jwt_analyzer", data)
+
+        if result.get("success"):
+            analysis = result.get("jwt_analysis_results", {})
+            vuln_count = len(analysis.get("vulnerabilities", []))
+            algorithm = analysis.get("token_info", {}).get("algorithm", "unknown")
+
+            logger.info(f"✅ JWT analysis completed: {vuln_count} vulnerabilities found")
+            logger.info(f"🔐 Token algorithm: {algorithm}")
+
+            if vuln_count > 0:
+                logger.warning(f"⚠️  Found {vuln_count} JWT vulnerabilities!")
+                for vuln in analysis.get("vulnerabilities", [])[:3]:  # Show first 3
+                    severity = vuln.get("severity", "UNKNOWN")
+                    vuln_type = vuln.get("type", "unknown")
+                    logger.warning(f"   ├─ [{severity}] {vuln_type}")
+        else:
+            logger.error("❌ JWT analysis failed")
+
+        return result
