+++ mcp_tools/crypto_attack/hashpump.py
@@ -1,0 +1,35 @@
+# mcp_tools/crypto_attack/hashpump.py
+
+from typing import Dict, Any
+
+def register_hashpump_tool(mcp, hexstrike_client, logger):
+
+    @mcp.tool()
+    def hashpump_attack(signature: str, data: str, key_length: str, append_data: str, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute HashPump for hash length extension attacks with enhanced logging.
+
+        Args:
+            signature: Original hash signature
+            data: Original data
+            key_length: Length of secret key
+            append_data: Data to append
+            additional_args: Additional HashPump arguments
+
+        Returns:
+            Hash length extension attack results
+        """
+        payload = {
+            "signature": signature,
+            "data": data,
+            "key_length": key_length,
+            "append_data": append_data,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔐 Starting HashPump attack")
+        result = hexstrike_client.safe_post("api/tools/hashpump", payload)
+        if result.get("success"):
+            logger.info(f"✅ HashPump attack completed")
+        else:
+            logger.error(f"❌ HashPump attack failed")
+        return result
