+++ mcp_tools/exploit_framework/pwninit.py
@@ -1,0 +1,36 @@
+# mcp_tools/exploit_framework/pwninit.py
+
+from typing import Dict, Any
+
+def register_pwninit_tool(mcp, hexstrike_client, logger):
+    
+    @mcp.tool()
+    def pwninit_setup(binary: str, libc: str = "", ld: str = "",
+                     template_type: str = "python", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute pwninit for CTF binary exploitation setup.
+
+        Args:
+            binary: Binary file to set up
+            libc: Libc file to use
+            ld: Loader file to use
+            template_type: Template type (python, c)
+            additional_args: Additional pwninit arguments
+
+        Returns:
+            CTF binary exploitation setup results
+        """
+        data = {
+            "binary": binary,
+            "libc": libc,
+            "ld": ld,
+            "template_type": template_type,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔧 Starting pwninit setup: {binary}")
+        result = hexstrike_client.safe_post("api/tools/pwninit", data)
+        if result.get("success"):
+            logger.info(f"✅ pwninit setup completed")
+        else:
+            logger.error(f"❌ pwninit setup failed")
+        return result
