+++ mcp_tools/smb_enum/__init__.py
@@ -1,0 +1,13 @@
+from .enum4linux import register_enum4linux_tool
+from .netexec import register_netexec_tool
+from .smbmap import register_smbmap_tool
+from .nbtscan import register_nbtscan_tool
+from .rpcclient import register_rpcclient_tool
+
+__all__ = [
+    "register_enum4linux_tool",
+    "register_netexec_tool",
+    "register_smbmap_tool",
+    "register_nbtscan_tool",
+    "register_rpcclient_tool"
+]
