+++ mcp_tools/web_fuzz/__init__.py
@@ -1,0 +1,17 @@
+from .dirb import register_dirb_tool
+from .ffuf import register_ffuf_tool
+from .dirsearch import register_dirsearch_tools
+from .gobuster import register_gobuster
+from .feroxbuster import register_feroxbuster_tool
+from .dotdotpwn import register_dotdotpwn_tool
+from .wfuzz import register_wfuzz_tool
+
+__all__ = [
+    "register_dirb_tool",
+    "register_ffuf_tool",
+    "register_dirsearch_tools",
+    "register_gobuster",
+    "register_feroxbuster_tool",
+    "register_dotdotpwn_tool",
+    "register_wfuzz_tool",
+]
