+++ mcp_tools/web_fuzz/feroxbuster.py
@@ -1,0 +1,33 @@
+# mcp_tools/web_fuzz/feroxbuster.py
+
+from typing import Dict, Any
+
+def register_feroxbuster_tool(mcp, hexstrike_client, logger):
+    
+    @mcp.tool()
+    def feroxbuster_scan(url: str, wordlist: str = "/usr/share/wordlists/dirb/common.txt", threads: int = 10, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Feroxbuster for recursive content discovery with enhanced logging.
+
+        Args:
+            url: The target URL
+            wordlist: Wordlist file to use
+            threads: Number of threads
+            additional_args: Additional Feroxbuster arguments
+
+        Returns:
+            Content discovery results
+        """
+        data = {
+            "url": url,
+            "wordlist": wordlist,
+            "threads": threads,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Feroxbuster scan: {url}")
+        result = hexstrike_client.safe_post("api/tools/feroxbuster", data)
+        if result.get("success"):
+            logger.info(f"✅ Feroxbuster scan completed for {url}")
+        else:
+            logger.error(f"❌ Feroxbuster scan failed for {url}")
+        return result
