+++ mcp_tools/web_fuzz/wfuzz.py
@@ -1,0 +1,31 @@
+# mcp_tools/web_fuzz/wfuzz.py
+
+from typing import Dict, Any
+
+def register_wfuzz_tool(mcp, hexstrike_client, logger):
+
+    @mcp.tool()
+    def wfuzz_scan(url: str, wordlist: str = "/usr/share/wordlists/dirb/common.txt", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Wfuzz for web application fuzzing with enhanced logging.
+
+        Args:
+            url: The target URL (use FUZZ where you want to inject payloads)
+            wordlist: Wordlist file to use
+            additional_args: Additional Wfuzz arguments
+
+        Returns:
+            Web application fuzzing results
+        """
+        data = {
+            "url": url,
+            "wordlist": wordlist,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Wfuzz scan: {url}")
+        result = hexstrike_client.safe_post("api/tools/wfuzz", data)
+        if result.get("success"):
+            logger.info(f"✅ Wfuzz scan completed for {url}")
+        else:
+            logger.error(f"❌ Wfuzz scan failed for {url}")
+        return result
