+++ mcp_tools/web_scan/xsser.py
@@ -1,0 +1,31 @@
+# mcp_tools/web_scan/xsser.py
+
+from typing import Dict, Any
+
+def register_xsser_tool(mcp, hexstrike_client, logger):
+    
+    @mcp.tool()
+    def xsser_scan(url: str, params: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute XSSer for XSS vulnerability testing with enhanced logging.
+
+        Args:
+            url: The target URL
+            params: Parameters to test
+            additional_args: Additional XSSer arguments
+
+        Returns:
+            XSS vulnerability test results
+        """
+        data = {
+            "url": url,
+            "params": params,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting XSSer scan: {url}")
+        result = hexstrike_client.safe_post("api/tools/xsser", data)
+        if result.get("success"):
+            logger.info(f"✅ XSSer scan completed for {url}")
+        else:
+            logger.error(f"❌ XSSer scan failed for {url}")
+        return result
